# public/openapi.yaml
openapi: 3.0.0
info:
  title: Google Cloud User Access Token Retrieval and GCP API
  version: 1.0.0
servers:
  - url: http://localhost:5000
paths:
  /idptoken:
    get:
      operationId: idptoken
      summary: Returns a new ID token form the IDP for the logged-in user.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  id_token:
                    type: string
                  error:
                    type: string
      tags:
        - Authentication
  /gcpaccesstoken:
    get:
      operationId: gcpaccesstoken
      summary: Exchanges an input token for a GCP access token.
      parameters:
        - name: input_token
          in: query
          description: The input token to exchange for a GCP access token.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  error:
                    type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
      tags:
        - GCP Access Token
  /gcpprojects:
    post:
      operationId: gcpprojects
      summary: Retrieves a list of GCP projects using the `resourcemanager.googleapis.com/projects` API.
      consumes:
        - application/json
      parameters:
        - in: body
          name: access_token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      type: string
                  error:
                    type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
      tags:
        - GCP Projects
  /securityconsult:
    post:
      operationId: securityconsult
      summary: Consult if a given request/query is secure and provide security expert advice.
      consumes:
        - application/json
      parameters:
        - in: body
          name: query_content
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items:
                      type: string
                  error:
                    type: string
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
      tags:
        - GCP Projects
  /bigquery:
    post:
      operationId: bigqueryDatasetsGet
      summary: Returns the dataset specified by projectID anddatasetID.
      parameters:
        - name: access_token
          in: body
          required: true
          schema:
            type: string
        - name: projectId
          in: body
          description: Required. Project ID of the requested dataset.
          required: true
          schema:
            type: string
        - name: datasetId
          in: body
          description: Required. Dataset ID of the requested dataset.
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dataset'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
      tags:
        - BigQuery
components:
  schemas:
    Dataset:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string